requirements.txt
pandas==2.2.2
openpyxl==3.1.5
sqlalchemy==2.0.30
mysql-connector-python==9.6.0
PyYAML==6.0.1
python-dotenv==1.0.1
fpdf2==2.8.2